chore: initial aem.js commit
