+++ .hlxignore
@@ -1,0 +1,7 @@
+.*
+*.md
+karma.config.js
+LICENSE
+package.json
+package-lock.json
+test/*
