+++ 404.html
@@ -11,7 +11,7 @@
   <meta property="og:title" content="Page not found">
   <script src="/hlx_statics/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
   <script type="module">
-    import { sampleRUM } from '/hlx_statics/scripts/lib-helix.js';
+    import { sampleRUM } from '/hlx_statics/scripts/aem.js';
 
     window.addEventListener('load', () => {
       if (document.referrer) {
