+++ hlx_statics/blocks/info-columns/info-columns.js
@@ -5,7 +5,7 @@ import {
   applyBkgColorOverride,
   applyWidthOverride,
 } from '../../scripts/lib-adobeio.js';
-import { decorateLightOrDark } from '../../scripts/lib-helix.js';
+import { decorateLightOrDark } from '../../scripts/aem.js';
 
 /**
  * decorates the info-columns
