+++ hlx_statics/scripts/delayed.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM } from './lib-helix.js';
+import { sampleRUM } from './aem.js';
 import {
   focusRing,
   isHlxPath,
